NNGraph interface and implement for CompileAndRuntime
