+++ oneflow/api/python/framework/nn_graph.cpp
@@ -16,7 +16,9 @@ limitations under the License.
 #include <pybind11/pybind11.h>
 #include <string>
 #include "oneflow/api/python/of_api_registry.h"
-#include "oneflow/core/framework/nn_graph_if.h"
+#include "oneflow/core/framework/nn_graph.h"
+#include "oneflow/core/job/runtime.h"
+#include "oneflow/core/register/blob.h"
 
 namespace py = pybind11;
 
@@ -25,6 +27,22 @@ ONEFLOW_API_PYBIND11_MODULE("", m) {
   using namespace oneflow;
   py::class_<NNGraph, std::shared_ptr<NNGraph>>(m, "NNGraph")
       .def(py::init<const std::string&>())
-      .def_property_readonly("name", &NNGraph::job_name);
+      .def_property_readonly("name", &NNGraph::job_name)
+      .def("register_input_op_names",
+           [](NNGraph& graph, const std::vector<std::string>& input_op_names) {
+             return graph.RegisterInputOpNames(input_op_names).GetOrThrow();
+           })
+      .def("register_output_op_names",
+           [](NNGraph& graph, const std::vector<std::string>& output_op_names) {
+             return graph.RegisterOutputOpNames(output_op_names).GetOrThrow();
+           })
+      .def("register_variable_op_names_and_tensors",
+           [](NNGraph& graph, const std::vector<std::string>& variable_op_names,
+              const std::vector<std::shared_ptr<one::Tensor>>& variable_tensors) {
+             return graph.RegisterVariableOpNamesAndTensors(variable_op_names, variable_tensors)
+                 .GetOrThrow();
+           })
+      .def("complie_and_runtime",
+           [](NNGraph& graph) { return graph.CompileAndRuntime().GetOrThrow(); });
 }
 }  // namespace oneflow
