+++ oneflow/core/job/plan_util.h
@@ -18,6 +18,7 @@ limitations under the License.
 
 #include <functional>
 #include "oneflow/core/job/plan.pb.h"
+#include "oneflow/core/job/job.pb.h"
 
 namespace oneflow {
 
@@ -30,6 +31,8 @@ struct PlanUtil {
   static void ToDotFile(const Plan& plan, const std::string& filepath);
   static std::function<RegstDescProto*(int64_t)> MakeMutRegstDesc4Id(Plan* plan);
   static void SetForceInplaceMemBlock(Plan* plan);
+  static void DumpCtrlRegstInfoToPlan(Plan* plan);
+  static void GenCollectiveBoxingPlan(Job* job, Plan* plan);
   static const oneflow::OpAttribute& GetOpAttribute(const Plan* plan, int64_t job_id,
                                                     const oneflow::KernelConf& kernel_conf);
 };
